chore: category endpoint and  get products by category
